Refactored progress tracking across Request subclasses to improve accuracy.
There were several issues with the previous implementation. The main problem was the upload tasks uses the same `Progress` property for the upload and download portions of the request. This could cause all sorts of strange progress reporting behavior. Additionally, there were no APIs that returned the `Progress` object directly and you couldn’t specify what queue the progress closures should be called on.

This commit addresses all these limitations. The `downloadProgress` APIs are available on the data, download and upload requests. The `uploadProgress` APIs are only available on upload requests.